requirements.txt
matplotlib
seaborn
scikit-learn
torch
torchvision
streamlit
pillow
numpy